+++ lib/src/data/data_source/local/db_adapter/comment_db_adapter.dart
@@ -10,4 +10,6 @@ abstract class CommentDbAdapter {
 
   Stream<List<CommentHiveEntity>> listenCommentEntities(
       RequestBuilder<GetCommentRequest> request);
+
+  List<CommentHiveEntity> getCommentEntities(RequestBuilder<GetCommentRequest> request);
 }