+++ lib/src/data/data_source/local/hive_entity/channel_hive_entity_20.dart
@@ -101,6 +101,7 @@ class ChannelHiveEntity extends EkoObject{
 
   ChannelHiveEntity({
     this.channelId,
+    this.defaultSubChannelId,
     this.isDistinct,
     this.metadata,
     this.type,